# lock by default - add safe_lock option for control