--- a/test/core.jl
+++ b/test/core.jl
@@ -95,4 +95,4 @@ p = Progress(10)
 @test p.safe_lock == (Threads.nthreads() > 1)
 p = Progress(10; safe_lock = false)
 @test p.safe_lock == false
-@test simple_sum(10; safe_lock = false) ≈ simple_sum(10; safe_lock = false)
+@test simple_sum(10; safe_lock = true) ≈ simple_sum(10; safe_lock = false)