This is another possible fix for: https://github.com/timholy/ProgressMeter.jl/issues/317 and https://github.com/timholy/ProgressMeter.jl/issues/232

Here we remove the automatic identification of threading, and always use a lock by default. An option `safe_lock` is added, which is by default `true` and must be set to `false` to disable the lock.

Relative to the current stable version I don't see any measurable performance difference:

Current stable:

```julia
julia> function prog_perf(n)
           prog = Progress(n)
           x = 0.0
           for i in 1:n
               x += rand()
               next!(prog)
           end
           return x
       end
prog_perf (generic function with 1 method)

julia> @time prog_perf(10^7)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:05
  5.938973 seconds (122.22 M allocations: 1.889 GiB, 2.62% gc time, 28.52% compilation time)
4.999510125125184e6

julia> @time prog_perf(10^7)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:03
  3.953561 seconds (119.76 M allocations: 1.785 GiB, 2.68% gc time)
4.999360964692622e6

julia> @time prog_perf(10^7)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:04
  4.044011 seconds (119.74 M allocations: 1.784 GiB, 2.60% gc time)
5.0003511864208765e6
```

With this PR:

```julia
julia> function prog_perf(n; safe_lock=true)
           prog = Progress(n; safe_lock)
           x = 0.0
           for i in 1:n
               x += rand()
               next!(prog)
           end
           return x
       end
prog_perf (generic function with 1 method)

julia> @time prog_perf(10^7)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:04
  4.068567 seconds (119.76 M allocations: 1.785 GiB, 4.76% gc time)
5.002101485254395e6

julia> @time prog_perf(10^7)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:03
  3.992197 seconds (119.74 M allocations: 1.784 GiB, 3.91% gc time)
4.999887783558271e6

julia> @time prog_perf(10^7; safe_lock=false)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:03
  3.704703 seconds (119.74 M allocations: 1.784 GiB, 4.53% gc time)
4.998432809492471e6

julia> @time prog_perf(10^7; safe_lock=false)
Progress: 100%|█████████████████████████████████████████████████████| Time: 0:00:03
  3.749583 seconds (119.73 M allocations: 1.784 GiB, 4.10% gc time)
5.000024621577769e6
```

There is an apparent small but consistent advantage when `safe_lock=false`, which is expected.