@lmiq 2024-07-07
Shoudn´t we release one of these fixes? None is breaking, and while the ˋsafe_lockˋ option is not documented, it can be reverted. If a better solution if proposed later, it can be implemented anyway.