@lmiq 2024-07-07
In https://github.com/timholy/ProgressMeter.jl/issues/232#issuecomment-1685158237 @IanButterworth mentions that the use of the progress meter slows down 9x the tight loop. I actually see a much greater slowdown with the released version, but my tests with this version, for which the lock can be removed, show that the slowdown is not really related to the lock.