@lmiq 2024-07-08
The main issue I see here is that the comparison with **not** using the progress meter is so great, that it makes me wander if this discussion about performance is really relevant. This is what I get here with the functions used in the test suite for this:

```
julia> @time noprog_perf(10^7)
  0.014616 seconds
5.001148339582725e6

julia> @time prog_perf(10^7)
Progress: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:05
  5.515208 seconds (119.79 M allocations: 1.785 GiB, 0.55% gc time)
5.000329702123166e6

julia>
```

This is with the current released version. Does it make sense to discuss the performance differences of using or not a lock in this context? Seems to me that using it for a tight loop simply does not make sense with the current status of things.